Update entry_main_program.c
I was unable to solve the memory. And or hard drive leaks as I began to have short term memory limitations. That is to say I could not live the interworkings of the program within my mind in order to fix it or change it.
Change to c++ and using creator environment