Add files via upload
This is the exact medium quality data is sent. This is always going to be combined with the low quality data. Medium only quality data is surprisingly unrecognizable and or makes the viewer believe that the sender is trying to be funny by using a filter, never does it look like a clear video when medium quality only