Add files via upload
This is the client view when the client has buffered low and medium quality video. It uses what can be considered as high fidelity data to fill in the gaps that are still missing from the high resolution pixels.

Personal double blind studies using a computer to give the test conclude that when low and medium photos are combined the viewer usually believe they are looking at the highest quality of the source video.